requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
Authlib==0.15.5
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
cryptography==36.0.1
deepdiff==5.7.0
discord.py==1.7.3
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
multidict==6.0.2
ordered-set==4.0.2
outcome==1.1.0
prompt-toolkit==3.0.26
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
rfc3986==1.5.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tda-api==1.5.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
websockets==10.1
wsproto==1.0.0
xmltodict==0.12.0
yarl==1.7.2
